Do not use minizinc driver but pipe flatzinc to solns2out directly.
This has the advantage that no shell is involved, so no subprocesses are created and killing the flatzinc interpreter works more reliably. Also, argument passing is done internally and does not suffer from the quoting problems in minizinc (e.g. filenames with spaces should work now). The compiled files are stored in a temporary directory, so we can run models that are read-only.

Needs thorough testing, probably still has some race conditions and not enough robustness against errors.