Highlighting protocol and html mode switching
1) Highlighting
  Highlighting urls take the form:
  "highlight?f1_1|sl1_1|sc1_1|el1_1|ec1_1|c;f1_2|sl1_2|sc1_2|el1_2|ec1_2|c;
            &f2|sl2|sc2|el2|ec2|c;

Where 'f' is a path to a file, 'sl', 'sc', 'el', 'ec' are the start line,
start column, end line and end column respectively, and c is an optional
field that is ignored (if the highlight is based on a mzn_path annotation
this will be the type of the expression pointed to)

Paths are separated by a '&' character. Each path will be assigned a
different colour by the highlighter.  Within a path, elements of the
path are separated by a semicolon ';' Each successive element in a path
is given the same colour but a different transparency.

2) mzn-html-start
  Additionally this commit adds support for the %%%mzn-html-start and
%%%mzn-html-end markers which allow solvers/tools to tell the IDE that
any output between these markers should be treated as html.