Fix signature for signal/slots using MessageWrapper
When connections are queued or across threads, Qt copies the arguments.
It uses the normalized signature [1] to determine the wanted type, which
has to made known via qRegisterType.

As only "MessageWrapper" is registered, but not "MessageWrapper&" (mind
the reference), this fails:

```
QObject::connect: Cannot queue arguments of type 'MessageWrapper&'
(Make sure 'MessageWrapper&' is registered using qRegisterMetaType().)
```

Use `const cpprofiler::MessageWrapper&` as it gets normalized to
`cpprofiler::MessageWrapper`, and using a const& avoids some copies.
In the header files use the fully qualified name, as this is what
Moc uses for creating the metaobject, this must match the type
in Q_DECLARE_METATYPE verbatim.

[1] https://doc.qt.io/qt-5/qmetaobject.html#normalizedSignature